requirements.txt
streamlit
python-docx
mistralai
python-dotenv	
pybase64
pillow	